LayoutTests/editing/pasteboard/data-transfer-get-data-non-normalized-types-expected.txt

Rich text
{
    "text/plain;": "",
    "text/uri-list;": "http://www.apple.com/",
    "text/html;": "<b style='color: green'></b>"
}
{
    "Text": "",
    "URL": "http://www.apple.com/",
    "text/html;charset=utf-8": "<b style='color: green'></b>"
}
